Time: 0 ms (100.00%) | Memory: 111.4 MB (11.10%) - LeetSync

+++ 2432-number-of-zero-filled-subarrays/number-of-zero-filled-subarrays.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     long long zeroFilledSubarray(vector<int>& nums) {
         long long count = 0;
-        long long count_val = 1;
+        int count_val = 1;
         int len = nums.size();
         for(int idx = 0 ; idx < len ; idx++){
             if(nums[idx] == 0){
